+++ src/main/java/com/parse/gwt/client/autobean/AutoBeanSerializer.java
@@ -8,25 +8,29 @@ import com.google.web.bindery.autobean.shared.AutoBeanUtils;
 
 public class AutoBeanSerializer {
 
-    protected AutoBeanFactory factory;
+    protected BaseFactory factory;
 
     public AutoBeanSerializer() {
         factory = null;
     }
 
-    public AutoBeanSerializer(AutoBeanFactory factory) {
+    public AutoBeanSerializer(BaseFactory factory) {
         super();
         this.factory = factory;
     }
 
     public <T> String encodeData(T data) {
         AutoBean<T> autoBean = AutoBeanUtils.getAutoBean(data);
-        return AutoBeanCodex.encode(autoBean).asString();
+        return AutoBeanCodex.encode(autoBean).getPayload();
     }
 
     public <T> T decodeData(Class<T> dataType, String json) {
         AutoBean<T> bean = AutoBeanCodex.decode(factory, dataType, json);
         return bean.as();
+    }
+
+    public <T> T make() {
+        return (T) this.factory.bean().as();
     }
 
 }