+++ src/main/java/com/parse/gwt/client/autobean/ParseFile.java
@@ -1,0 +1,10 @@
+package com.parse.gwt.client.autobean;
+
+public interface ParseFile {
+    String get__type();
+    String getClassName();
+    String getObjectId();
+    void set__type(String __type);
+    void getClassName(String className);
+    void setObjectId(String objectId);
+}